Create Makeshift_webhook.TS
this is a makeshift code that acts as webhook. it receives the incoming JSON data from the NextLegApi and saves it in a google sheet file, as csv. later to be extracted as a JSON by another script for the plugin.

NOTE: do not use this implementation in your final product, might not be scalable, this code is to just give you an idea on how the supporting features for the plugin works and the type of data format that the bubble plugin is expecting.